Update PlayerMixin.java
Instead of hooking to LocalPlayer.class, the mod now hook to LivingEntity.class, and inject to setHealth(). The downside is this method will fire everytime an entity is changed it's HP (both healing and damaging). So we check if this is the LocalPlayer, and because lowest healing in Mineraft is 1HP, we skip the check if it's heal or damage.